requirements.txt
asgiref==3.5.2
atomicwrites==1.4.0
attrs==21.4.0
autopep8==1.6.0
colorama==0.4.4
Django==3.1.13
django-factory==0.11
djangorestframework==3.13.1
extras==1.0.0
factory-boy==3.2.1
Faker==13.12.1
fixtures==4.0.0
iniconfig==1.1.1
packaging==21.3
pbr==5.9.0
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
testtools==2.5.0
toml==0.10.2
tomli==2.0.1
